Replaces functional tests of importer form to just the download template link
